refactor: move and rename files related to Geo and update exports
