+++ giraffe/src/components/geo/index.ts
@@ -1,0 +1,1 @@
+export {GeoLayer} from './GeoLayer'
