+++ giraffe/src/utils/geo.ts
@@ -1,14 +1,15 @@
 import L from 'leaflet'
-import {GeoViewLayer} from '..'
-import {GeoTable} from '../components/geo/processing/GeoTable'
 import {
   ClusterAggregation,
   GeoCircleViewLayer,
   GeoPointMapViewLayer,
   GeoQueryVariables,
-} from '../types/geo'
-import {formatValue} from '../components/geo/processing/toolTips'
-import {getColor} from '../components/geo/dimensionCalculations'
+  GeoViewLayer,
+} from '../types'
+import {GeoTable} from '../components/Geo/processing/GeoTable'
+
+import {formatValue} from '../components/Geo/processing/toolTips'
+import {getColor} from '../components/Geo/dimensionCalculations'
 
 // Constants
 export const ZOOM_FRACTION = 8
